+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## NEW FEATURES
 * Add support for images with colormaps (also known as lookup tables (LUTs)).
+* Add a print method for `ijtiff_img`s.
 
 
 # `ijtiff` 2.0.5